pyproject.toml
[build-system]
requires = ["setuptools", "setuptools_scm"]
build-backend = "setuptools.build_meta"

[project]
name = "ts_ess_controller"
description = "Rubin Observatory Environment Sensors Support"
license = { text = "GPL" }
classifiers = ["Programming Language :: Python :: 3"]
urls = { documentation = "https://ts-ess-controller.lsst.io", source_code = "https://github.com/lsst-ts/ts_ess_controller" }
dynamic = ["version"]
dependencies = ["pyserial", "pylibftdi", "aioserial"]

[tool.setuptools.dynamic]
version = { attr = "setuptools_scm.get_version" }

[tool.setuptools.packages.find]
where = ["python"]

[project.scripts]
run_ess_controller = "lsst.ts.ess.controller:run_ess_controller"

[tool.setuptools_scm]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[project.optional-dependencies]
dev = ["documenteer[pipelines]"]
